Remove unnecessary tests in CompareV2.
